+++ supramatch/discovery/__init__.py
@@ -1,0 +1,7 @@
+"""
+External chemical-data acquisition for supramatch.
+
+Modules:
+    pubchem_client: Compound discovery via PubChem PUG REST
+    price_lookup: Vendor price quotes via ChemPrice (mcule, molport)
+"""